requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
atari-py==0.2.9
box2d-py==2.3.8
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==1.2.1
contourpy==1.3.0
cvxopt==1.3.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
dm-tree==0.1.8
exceptiongroup==1.2.2
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.56.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
gym==0.15.7
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.2
joblib==1.4.2
keras==3.0.5
kiwisolver==1.4.7
libclang==18.1.1
lxml==4.9.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.3.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpi4py==4.0.3
namex==0.0.8
numpy==1.26.4
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==1.5.3
pandas-datareader==0.10.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
Pillow==9.5.0
plotly==4.14.3
pluggy==1.5.0
pmdarima==2.0.4
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyglet==1.5.0
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
rich==13.9.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.10.1
setuptools-scm==8.2.0
six==1.17.0
# -e git+https://github.com/openai/spinningup.git@8317a49e4ccd4fac98a07307d1d787df9f1775e3#egg=spinup
stack-data==0.6.3
statsmodels==0.13.5
swig==4.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==2.5.0
threadpoolctl==3.5.0
tomli==2.2.1
torch==1.13.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
universal-portfolios==0.4.12
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
